Added generic api controller
